fix AlgoliaReindexAllJob to batch into steps
Fix to use jobData so that it is possible to run over multiple steps & if the job dies or runs out of memory etc.
Complete overhaul of this class to actually do what it says in the setup description - do the heavy lifting / batching at setup and then step through them 1 chunk at a time using QueuedJob's currentStep.